new versions of python and html files, CORS changes

+++ record_rest_srv_cors.py
@@ -5,7 +5,7 @@
 #
 # David O'Connell
 
-from flask import Flask, jsonify, url_for, request, redirect, abort
+from flask import Flask, jsonify, url_for, request, redirect, abort, make_response
 from flask_cors import CORS, cross_origin
 #from recordDAOframework import recordDAO
 from recordDAO import recordDAO
@@ -17,6 +17,13 @@ app = Flask(__name__, static_url_path='', static_folder='staticpages')
 cors = CORS(app) # only needed if you want to expose all routes to cross origin
 app.config['CORS_HEADERS'] = 'Content-Type' # not sure if needed
 
+def _build_cors_preflight_response():
+    response = make_response()
+    response.headers.add("Access-Control-Allow-Origin", "*")
+    response.headers.add("Access-Control-Allow-Headers", "*")
+    response.headers.add("Access-Control-Allow-Methods", "*")
+    return response
+
 #------------------------------------------------------------------------
 # Index
 
@@ -27,28 +34,45 @@ def index():
 
 #------------------------------------------------------------------------
 # Get all records
+# curl -X GET http://127.0.0.1:5000/records
 
-@app.route('/records', methods=['GET'])
+@app.route('/records', methods=["GET", "OPTIONS"])
 @cross_origin()
 
 def get_all():
+    if request.method == "OPTIONS": # CORS preflight
+        print("get_all OPTIONS")
+        return _build_cors_preflight_response()
     # Get the response object
-    return jsonify(recordDAO.get_all_records())
+    #return jsonify(recordDAO.get_all_records())
+    response = jsonify(recordDAO.get_all_records())
+    #response.headers.add("Access-Control-Allow-Origin", "*")
+    return response
 
 #------------------------------------------------------------------------
 # Get record by ID
+# curl -X GET http://127.0.0.1:5000/records/2
 
-@app.route('/records/<int:id>', methods=['GET'])
+@app.route('/records/<int:id>', methods=["GET", "OPTIONS"])
 @cross_origin()
 
 def find_by_id(id):
+    if request.method == "OPTIONS": # CORS preflight
+        print("find_by_id OPTIONS")
+        return _build_cors_preflight_response()
     # Get the response object
-    return jsonify(recordDAO.find_record_by_id(id))
+    #return jsonify(recordDAO.find_record_by_id(id))
+    response = jsonify(recordDAO.find_record_by_id(id))
+    #response.headers.add("Access-Control-Allow-Origin", "*")
+    return response
 
 #------------------------------------------------------------------------
 # Create a record
+#curl -X POST -H "Content-Type: application/json" -d "{\"title\":\"Closer\", 
+# \"artist\":\"Joy Division\", \"year\":1981, \"genre\":\"post punk\"}" 
+# http://127.0.0.1:5000/records
 
-@app.route('/records', methods=['POST'])
+@app.route('/records', methods=["POST", "OPTIONS"])
 @cross_origin()
 
 def create():
@@ -56,6 +80,10 @@ def create():
     record = {}
     record_string = request.json
     
+    if request.method == "OPTIONS": # CORS preflight
+        print("create OPTIONS")
+        return _build_cors_preflight_response()
+
     # build the record to create - check each field is present
     if "title" not in record_string:
         abort(400)
@@ -69,21 +97,30 @@ def create():
         abort(400)
     record["year"] = record_string["year"]
 
+
     if "genre" not in record_string:
         abort(400)
     record["genre"] = record_string["genre"]
 
     # Create the record, get the response object
     print("Creating", record)
-    return jsonify(recordDAO.create_record(record))
+    #return jsonify(recordDAO.create_record(record))
+    response = jsonify(recordDAO.create_record(record))
+    #response.headers.add("Access-Control-Allow-Origin", "*")
+    return response
 
 #------------------------------------------------------------------------
 # Update a record
+#curl -X PUT -H "Content-Type: application/json" -d "{\"title\":\"Further\"}" 
+#http://127.0.0.1:5000/records/5
 
-@app.route('/records/<int:id>', methods=['PUT'])
+@app.route('/records/<int:id>', methods=["PUT", "OPTIONS"])
 @cross_origin()
 
 def update(id):
+    if request.method == "OPTIONS": # CORS preflight
+        return _build_cors_preflight_response()
+
     # Read the JSON from the body of the incoming request
     record = {}
     record_string = request.json
@@ -100,19 +137,29 @@ def update(id):
 
     # Update the record, get the response object
     print("Updating record", id, "to", record)
-    return jsonify(recordDAO.update_record(id, record))
+    #return jsonify(recordDAO.update_record(id, record))
+    response = jsonify(recordDAO.update_record(id, record))
+    #response.headers.add("Access-Control-Allow-Origin", "*")
+    return response
 
 #------------------------------------------------------------------------
 # Delete a record
+# curl -X DELETE http://127.0.0.1:5000/records/5
 
-@app.route('/records/<int:id>', methods=['DELETE'])
+@app.route('/records/<int:id>', methods=["DELETE", "OPTIONS"])
 @cross_origin()
 
 def delete(id):
+    if request.method == "OPTIONS": # CORS preflight
+        return _build_cors_preflight_response()
     # Delete the record, get the response object
     if recordDAO.delete_record(id):
-        return jsonify({"done":True})
         #return jsonify(recordDAO.delete_record(id))
+        #return jsonify({"done":True})
+        response = jsonify({"done":True})
+        #response.headers.add("Access-Control-Allow-Origin", "*")
+        return response
+    
 
 if __name__ == "__main__":
     app.run(debug=True)
